refactor: short 타입 수

+++ chat/src/main/java/doktoribackend/quiz/domain/QuizChoiceId.java
@@ -12,5 +12,5 @@ import java.io.Serializable;
 public class QuizChoiceId implements Serializable {
 
     private Long roomId;
-    private Short choiceNumber;
+    private Integer choiceNumber;
 }
